requirements.txt
lxml
click
pyyaml
pytest
